a few tweaks
removed the beacon from secoff spawn loadouts because they have gun cases in their lockers, and we don't Need More Guns.

Also tweaked Cargodise Lost a bit to remove some of the Ultra Gamer Loot and changed a bit of flavor text and role titles to make more sense for an independent freighter.